requirements.txt
pdfplumber==0.11.4
PyPDF2==3.0.1
tabula-py==2.9.3
pandas==2.2.2
openpyxl==3.1.5
regex==2024.7.24
python-dateutil==2.9.0
camelot-py[cv]==0.11.0
pdfminer.six==20231228
reportlab==4.2.0
